refactor(web): normalize route-local folder structure (_components) and resolve duplicate component names
