+++ apps/web/src/app/[locale]/event/[eventCode]/page.tsx
@@ -1,6 +1,6 @@
 import { notFound } from 'next/navigation';
 
-import { Footer } from '../../landing/Footer';
+import { Footer } from '../../landing/_components/Footer';
 import { ChristmasContent } from '../(christmas)';
 import { EventEndOverlay } from '../(common)/EventEndOverlay';
 import { HalloweenContent } from '../(halloween)/Content';
