+++ apps/web/src/app/[locale]/landing/_components/AvailablePetSection/AnimalCardList.tsx
