+++ apps/web/src/app/[locale]/landing/_components/AvailablePetSection/AvailablePetSection.tsx
