+++ apps/web/src/app/[locale]/landing/_components/ChoosePetSection/ChoosePetSection.tsx
