+++ apps/web/src/app/[locale]/landing/_components/ChoosePetSection/index.ts
