+++ apps/web/src/app/[locale]/landing/_components/Footer/Footer.tsx
