+++ apps/web/src/app/[locale]/landing/_components/MainSection/MainSection.tsx
