+++ apps/web/src/app/[locale]/landing/_components/MainSection/MainSlider.tsx
