+++ apps/web/src/app/[locale]/landing/_components/MainSection/index.ts
