+++ apps/web/src/app/[locale]/landing/_components/RankingSection/GameConsole/Joystick.tsx
