+++ apps/web/src/app/[locale]/landing/_components/RankingSection/GameConsole/PixelNoiseEffect.tsx
