+++ apps/web/src/app/[locale]/landing/_components/RankingSection/MobileGameConsole/MobileGameConsole.tsx
