+++ apps/web/src/app/[locale]/landing/_components/RankingSection/MobileRankingTable.tsx
