+++ apps/web/src/app/[locale]/landing/_components/RankingSection/RankingTable.tsx
