+++ apps/web/src/app/[locale]/landing/index.ts
@@ -1,8 +1,8 @@
-import { RankingServerSide } from './RankingSection/RankingServerSide';
-import { AvailablePetSection } from './AvailablePetSection';
-import { ChoosePetSection } from './ChoosePetSection';
-import { Footer } from './Footer';
-import { HavePetWaySection } from './HavePetWaySection';
-import { MainSection } from './MainSection';
+import { RankingServerSide } from './_components/RankingSection/RankingServerSide';
+import { AvailablePetSection } from './_components/AvailablePetSection';
+import { ChoosePetSection } from './_components/ChoosePetSection';
+import { Footer } from './_components/Footer';
+import { HavePetWaySection } from './_components/HavePetWaySection';
+import { MainSection } from './_components/MainSection';
 
 export { AvailablePetSection, ChoosePetSection, Footer, HavePetWaySection, MainSection, RankingServerSide };
