+++ apps/web/src/components/ConfirmDialog/index.ts
@@ -1,0 +1,1 @@
+export * from './ConfirmDialog';
